Crate hua_core

Modules

extra

Structs

ComponentPaths

DependencyGraph

Downloader

An Downloader to retrieve files from different sources

Generation

GenerationManager

Package

Recipe

A Recipe to build an Package from.

Requirement

Store

A Store that contains all the packages installed by any user

User

A single User who owns several generations.

UserManager

Manages all users.

Version

SemVer version as defined by https://semver.org.

VersionReq

SemVer version requirement describing the intersection of some version comparators, such as >=1.2.3, <1.8.

Enums

Binary

Component

Conflict

DependencyError

Error

Step